Drop 64 bit target and a quick model test
Droping the 64 bit target and switch to 32 bit because the BPL wish to be a 32 bit target. A quick test of the components with the TDocument I made a quick test with Data components described in chapters 4.5.3 and 4.5.4. Look good... no compilation and runtime error and I see the columns in the grid.